% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscc.R
\name{get_pscc}
\alias{get_pscc}
\title{Philippine Standard Commodity Classification (PSCC)}
\usage{
get_pscc(
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PSCC to retrieve. Default is "2022".}

\item{level}{Character. Level of detail to retrieve. Options are "all", "sections", "chapters", "headings", "commodity". Default is "all".}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A tibble containing the requested PSCC data.
}
\description{
Philippine Standard Commodity Classification (PSCC)
}
\examples{
\dontrun{
get_pscc(token = "your_api_token")
}
}
