% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psccs.R
\name{get_psccs}
\alias{get_psccs}
\title{Philippine Standard Classification of Crime for Statistical Purposes (PSCCS)}
\usage{
get_psccs(
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PSCCS to retrieve. Default is "2018".}

\item{level}{Character. Level of detail to retrieve. Options are "all", "section", "divisions", "groups", "classes", "sub-classes". Default is "all".}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A tibble containing the requested PSCCS data.
}
\description{
Philippine Standard Classification of Crime for Statistical Purposes (PSCCS)
}
\examples{
\dontrun{
get_psccs(token = "your_api_token")
}
}
