% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcoicop.R
\name{get_pcoicop}
\alias{get_pcoicop}
\title{Philippine Classification of Individual Consumption According to Purpose (PCOICOP)}
\usage{
get_pcoicop(
  token = NULL,
  version = "2020",
  level = "all",
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PCOICOP to retrieve. Default is "2020".}

\item{level}{Character. Level of detail to retrieve. Options are "all", "divisions", "groups", "class", "sub-class", "item", "subitem". Default is "all".}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A tibble containing the requested PCOICOP data.
}
\description{
Philippine Classification of Individual Consumption According to Purpose (PCOICOP)
}
\examples{
\dontrun{
get_pcoicop(token = "your_api_token")
}
}
