% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcpc.R
\name{get_pcpc}
\alias{get_pcpc}
\title{Philippine Central Product Classification (PCPC)}
\usage{
get_pcpc(
  ...,
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PCPC to retrieve. Default is "2002".}

\item{level}{Character. Level of detail to retrieve. Options are "all", "sections", "divisions", "groups", "classes", "sub-classes", "item". Default is "all".}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A tibble containing the requested PCPC data.
}
\description{
Philippine Central Product Classification (PCPC)
}
\examples{
\dontrun{
get_pcpc(token = "your_api_token")
}
}
