% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psced.R
\name{get_psced}
\alias{get_psced}
\title{Retrieve Philippine Standard Classification of Education (PSCED)}
\usage{
get_psced(
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PSCED dataset. Default is \code{"2017"}.}

\item{level}{Character. Classification level. Default is \code{"detailedfield"}.}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A data frame of PSCED classifications.
}
\description{
Retrieve Philippine Standard Classification of Education (PSCED)
}
\examples{
\dontrun{
get_psced(token = "your_api_token")
}
}
