% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psic.R
\name{get_psic}
\alias{get_psic}
\title{Retrieve Philippine Standard Industrial Classification (PSIC)}
\usage{
get_psic(
  ...,
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PSIC dataset. Default is \code{"2019"}.}

\item{level}{Character. Classification level (e.g., \code{"sub-classes"}).}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A data frame of PSIC classifications.
}
\description{
Retrieve Philippine Standard Industrial Classification (PSIC)
}
\examples{
\dontrun{
get_psic(token = "your_api_token")
}
}
