% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psoc.R
\name{get_psoc}
\alias{get_psoc}
\title{Retrieve Philippine Standard Occupational Classification (PSOC)}
\usage{
get_psoc(
  token = NULL,
  version = NULL,
  level = NULL,
  harmonize = TRUE,
  minimal = TRUE,
  cols = NULL
)
}
\arguments{
\item{token}{Character. API access token.}

\item{version}{Character. Version of the PSOC dataset. Default is \code{"2012"}.}

\item{level}{Character. Classification level (e.g., \code{"unit"}).}

\item{harmonize}{Logical. If \code{TRUE}, formats and standardizes the returned data. Default is \code{TRUE}.}

\item{minimal}{Logical. If \code{TRUE}, returns a simplified dataset. Default is \code{TRUE}.}

\item{cols}{Optional. Character vector of additional columns to include when \code{minimal = FALSE}.}
}
\value{
A data frame of PSOC classifications.
}
\description{
Retrieve Philippine Standard Occupational Classification (PSOC)
}
\examples{
\dontrun{
get_psoc(token = "your_api_token")
}
}
